ver 5.9.2
Изменение главной страницы, авторизация, фильтры и разные баги и фичи
