+++ footer.php
@@ -28,23 +28,23 @@ echo '<div style="text-align: center; margin-top:50px; padding-top:20px; color: 
         </script>
         <!--<noscript><div><img src="//mc.yandex.ru/watch/23700652" style="position:absolute; left:-9999px;" alt="" /></div></noscript>-->
         <!-- /Yandex.Metrika counter -->
-        
+
         <!-- BEGIN JIVOSITE CODE {literal} -->
         <!--
-        <script type='text/javascript'>            
-            (function(){ 
+        <script type='text/javascript'>
+            (function(){
                 var widget_id = 'LsvANjk4FT';
                 var d=document;
                 var w=window;
                 function l(){
-                    var s = document.createElement('script'); 
-                    s.type = 'text/javascript'; 
-                    s.async = true; 
-                    s.src = '//code.jivosite.com/script/widget/'+widget_id; 
-                    var ss = document.getElementsByTagName('script')[0]; 
+                    var s = document.createElement('script');
+                    s.type = 'text/javascript';
+                    s.async = true;
+                    s.src = '//code.jivosite.com/script/widget/'+widget_id;
+                    var ss = document.getElementsByTagName('script')[0];
                     ss.parentNode.insertBefore(s, ss);
                 }
-                
+
                 if(d.readyState=='complete'){
                     l();
                 }
