+++ header.php
@@ -24,7 +24,7 @@ include_once "preheader.php";
     <link href="css/bootstrap-responsive.min.css" rel="stylesheet" />
     <link href="css/bootstrap-modal.css" rel="stylesheet" />
     <link href="css/bootstrap-datepicker.min.css" rel="stylesheet" />
-    <link href="css/style.css?v19" rel="stylesheet" />
+    <link href="css/style.css?v21" rel="stylesheet" />
 
     <script src="js/jquery.min.js" type="text/javascript"></script>
     <script src="js/bootstrap.min.js" type="text/javascript"></script>
@@ -33,7 +33,7 @@ include_once "preheader.php";
     <script src="js/jquery.mockjax.js" type="text/javascript"></script>
     <script src="js/jquery.autocomplete.js" type="text/javascript"></script> <!-- https://github.com/devbridge/jQuery-Autocomplete -->
     <script src="js/jquery.maskedinput.min.js" type="text/javascript"></script>
-    <script src="js/script.js?v43" type="text/javascript"></script>
+    <script src="js/script.js?v45" type="text/javascript"></script>
     <script src="js/jquery-textrange.js" type="text/javascript"></script>
     <script src="js/bootstrap-datepicker.min.js" type="text/javascript"></script>
     <script src="js/bootstrap-datepicker.ru.js" type="text/javascript"></script>
