+++ js/reg.js
@@ -1,3 +1,16 @@
-
   $('#emBirthdateLabelSup').html('Дата рождения<sup>*</sup>');
   $('.emBirthdate').attr('valid', 'required');
+// check date fields
+  $('#modalEditMember').on('show',  function() {
+    setTimeout(function () {
+      var form = $('#modalEditMember');
+      var a = parseDDMM (form.find(".emArrDate").val());
+      var b = parseDDMM (form.find(".emDepDate").val());
+        if(!$('.emArrDate').parent().hasClass('error') && !a){
+          $('.emArrDate').parent().addClass('error');
+        };
+        if(!$('.emDepDate').parent().hasClass('error') && !b){
+          $('.emDepDate').parent().addClass('error');
+      };
+  }, 1000);
+  });
